Move deprecated colour tokens to deprecated exports
