+++ libs/@guardian/source-foundations/src/deprecated-exports.ts
@@ -1,0 +1,38 @@
+/**
+ * DEPRECATED EXPORTS
+ *
+ * These will be removed in a future major release and should not be used.
+ */
+
+/* @deprecated Colours should now be imported from the `palette` object */
+export {
+	background,
+	brandBackground,
+	brandAltBackground,
+	border,
+	brandBorder,
+	brandAltBorder,
+	line,
+	brandLine,
+	brandAltLine,
+	text,
+	brandText,
+	brandAltText,
+} from './_deprecated/palette';
+
+/* @deprecated Colours should now be imported from the `palette` object */
+export {
+	brand,
+	brandAlt,
+	neutral,
+	error,
+	success,
+	news,
+	opinion,
+	sport,
+	culture,
+	lifestyle,
+	labs,
+	specialReport,
+	focus,
+} from './_deprecated/palette';
